#1934 Move channel members types in features/channel-members/types
