+++ twake/frontend/src/app/features/channel-members/types/channel-member-types.ts
@@ -1,0 +1,13 @@
+export type ChannelMemberTypeType = 'member' | 'guest' | 'bot';
+
+export type ChannelMemberNotificationLevel = 'all' | 'none' | 'mentions' | 'me';
+
+export type ChannelMemberType = {
+  user_id?: string;
+  channel_id?: string;
+  type?: ChannelMemberTypeType;
+  last_access?: number; //Timestamp in seconds
+  last_increment?: number; //Number
+  favorite?: boolean; //Did the user add this channel to its favorites
+  notification_level?: ChannelMemberNotificationLevel;
+};
