+++ twake/frontend/src/app/features/channels/api/channels-mine-api-client.ts
@@ -1,5 +1,5 @@
 import Api from '../../global/framework/api-service';
-import { ChannelMemberType, ChannelType } from 'app/features/channels/types/channel';
+import { ChannelType } from 'app/features/channels/types/channel';
 import { TwakeService } from '../../global/framework/registry-decorator-service';
 import { WebsocketRoom } from '../../global/types/websocket-types';
