+++ twake/frontend/src/app/features/channels/types/channel.ts
@@ -1,17 +1,5 @@
 import { Resource } from 'app/deprecated/CollectionsReact/Collections';
-
-export type ChannelMemberTypeType = 'member' | 'guest' | 'bot';
-export type ChannelMemberNotificationLevel = 'all' | 'none' | 'mentions' | 'me';
-
-export type ChannelMemberType = {
-  user_id?: string;
-  channel_id?: string;
-  type?: ChannelMemberTypeType;
-  last_access?: number; //Timestamp in seconds
-  last_increment?: number; //Number
-  favorite?: boolean; //Did the user add this channel to its favorites
-  notification_level?: ChannelMemberNotificationLevel;
-};
+import { ChannelMemberType } from 'app/features/channel-members/types/channel-member-types';
 
 export type ChannelType = {
   company_id?: string;
